+++ src/ttables.h
@@ -82,7 +82,7 @@ struct ContextVector{
 
 class TTable {
 	public:
-		TTable(int _n, int p_threshold);
+		TTable(int _n, double p_threshold);
 		double prob(const WordVector& e, const WordID& f);
 		double backoffProb(const WordVector& e, const WordID& f);
 		void Increment(const WordVector& e, const int& f);
@@ -103,7 +103,7 @@ class TTable {
 		void _ShowCounts(int index,Dict& d);
 		void _ShowTTable(int index,Dict& d);
 		void prune();
-		int prune_threshold;
+		double prune_threshold;
 	public:
 		VWV2WD ttables;
 		VWV2WD counts;
